* Added detectOS to be run before everything so that the CLI can use this function.

+++ scripts/shutdown.sh
@@ -25,27 +25,29 @@ dockerDownShutdown()
     local app_name="$1"
     local type="$2"
 
-    if [[ "$type" == "" ]]; then
-        # Used for standard app shutdown
-        if [[ $CFG_DOCKER_INSTALL_TYPE == "rootless" ]]; then
-            local result=$(runCommandForDockerInstallUser "cd $containers_dir$app_name && docker-compose down")
-            isSuccessful "Shutting down container for $app_name"
-            dockerDownShutdownSuccessMessage $app_name;
-        elif [[ $CFG_DOCKER_INSTALL_TYPE == "rooted" ]]; then
-            local result=$(cd $containers_dir$app_name && sudo docker-compose down)
-            isSuccessful "Shutting down container for $app_name"
-            dockerDownShutdownSuccessMessage $app_name;
-        fi
-    else
-        # Used for Shutting down Rooted type switcher
-        if [[ $type == "rootless" ]]; then
-            local result=$(runCommandForDockerInstallUser "cd $containers_dir$app_name && docker-compose down")
-            isSuccessful "Shutting down container for $app_name"
-            dockerDownShutdownSuccessMessage $app_name;
-        elif [[ $type == "rooted" ]]; then
-            local result=$(cd $containers_dir$app_name && sudo docker-compose down)
-            isSuccessful "Shutting down container for $app_name"
-            dockerDownShutdownSuccessMessage $app_name;
+    if [[ "$OS" == [1234567] ]]; then
+        if [[ "$type" == "" ]]; then
+            # Used for standard app shutdown
+            if [[ $CFG_DOCKER_INSTALL_TYPE == "rootless" ]]; then
+                local result=$(runCommandForDockerInstallUser "cd $containers_dir$app_name && docker-compose down")
+                isSuccessful "Shutting down container for $app_name"
+                dockerDownShutdownSuccessMessage $app_name;
+            elif [[ $CFG_DOCKER_INSTALL_TYPE == "rooted" ]]; then
+                local result=$(cd $containers_dir$app_name && sudo docker-compose down)
+                isSuccessful "Shutting down container for $app_name"
+                dockerDownShutdownSuccessMessage $app_name;
+            fi
+        else
+            # Used for Shutting down Rooted type switcher
+            if [[ $type == "rootless" ]]; then
+                local result=$(runCommandForDockerInstallUser "cd $containers_dir$app_name && docker-compose down")
+                isSuccessful "Shutting down container for $app_name"
+                dockerDownShutdownSuccessMessage $app_name;
+            elif [[ $type == "rooted" ]]; then
+                local result=$(cd $containers_dir$app_name && sudo docker-compose down)
+                isSuccessful "Shutting down container for $app_name"
+                dockerDownShutdownSuccessMessage $app_name;
+            fi
         fi
     fi
 }
